Implementing HabitEvent Logic
