+++ GittersAndSitters/app/src/main/res/layout/activity_habit_event.xml
@@ -2,7 +2,7 @@
 <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:id="@+id/activity_habit_event_history_id"
+    android:id="@+id/activity_habit_event_id"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical">
